Added charm system and save/load logic to the project.
Charms are meant to be passive abilities that the player can equip/remove and it gives them passive stats.

Save loading is working.

To add more save/load logic you must create a class in SaveProfile.cs which is derived from SaveProfileData. (See SaveCharmData for an example).

Save/Load is not yet testing in a shipping build and it must be due to using paths for scriptableobjects references.